Add worker heart

+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/entity/Worker.java
@@ -74,7 +74,7 @@ public class Worker {
      * Last heartbeat time
      */
     @Column(name = "last_heartbeat_time")
-    private Integer lastHeartbeatTime;
+    private Long lastHeartbeatTime;
 
     /**
      * Status
@@ -92,11 +92,11 @@ public class Worker {
      * Create time
      */
     @Column(name = "create_time")
-    private Integer createTime;
+    private Long createTime;
 
     /**
      * Update time
      */
     @Column(name = "update_time")
-    private Integer updateTime;
+    private Long updateTime;
 }
